[level 1] Title: 3진법 뒤집기, Time: 0.04 ms, Memory: 74.7 MB -BaekjoonHub

+++ 프로그래머스/1/68935. 3진법 뒤집기/3진법 뒤집기.java
@@ -1,12 +1,11 @@
-import java.util.*;
-
 class Solution {
+
     public int solution(int n) {
-        String base3 = Long.toString(n, 3);
+        String base3 = Integer.toString(n, 3);
 
         String reversedBase3 = new StringBuilder(base3)
                 .reverse().toString();
 
-        return (int) Long.parseLong(reversedBase3, 3);
+        return Integer.parseInt(reversedBase3, 3);
     }
 }
